Added trend holder
Added trend holder that holds the current price trend and updates it periodically in server side. This can be seen in action when clicking buy crypto the price of cryptos can be seen changed as the app sends new price request to the server. Updated buyviewmodel to throw null pointer exception when trying to select data range in combobox without selecting a crypto.